feat: smooth update progress with spring animation and compact dialog
- Replace shadcn/ui Progress with animate-ui Progress (spring physics)
- Use CountingNumber via ProgressValue for smooth percentage display
- Extract download logic into useUpdateDownload custom hook
- Add NaN guard when contentLength is unknown
- Reduce dialog width from 1200px fixed to max-w-4xl (896px)
